feat: refactor Home page and layout components
Add default Home page and remove Middleware redirection from root to /app
directory.

Restructure layout components for improved reusability across different layouts.
