+++ components/layouts/layout/logo.tsx
@@ -2,7 +2,7 @@ import { PrefetchRouterButton } from '@buttons/button/prefetchRouterButton';
 import { NetworkStatus } from '@components/notifications/networkStatus';
 import { Fragment as LogoContainerFragment, Fragment as LogoFragment } from 'react';
 
-export const LayoutLogo = () => {
+export const Logo = () => {
   return (
     <nav className='flex h-0 flex-row items-center'>
       <LogoContainerFragment>
