+++ pages/index.tsx
@@ -1,0 +1,11 @@
+import { Fragment } from 'react';
+
+const Home = () => {
+  return (
+    <Fragment>
+      <div>Home</div>
+    </Fragment>
+  );
+};
+
+export default Home;
